// framework/src/interfaces/SmartObjects/ICustomControl.ts
/*
 * File generated by Interface generator (dotup.dotup-vscode-interface-generator)
 * Date: 2022-01-13 04:25:56 
*/
import { ISmartObjectItem } from "./ISmartObjectItem";

export interface ICustomControlSmartObject {
    System_Name: string;
    Name: string;
    Description: string;
    Settings: string;
    SettingsAsObject: any;
    Namespace: string;
    Icon: string;
    Sample_Screenshot: string;
    Instructions: string;
    Help_URL: string;
    Target_Control_Types: string;
    Enabled: boolean;
    Javascript_Content: string;
    Javascript_URL: string;
    CSS_Content: string;
    CSS_URL: string;
    Extension: string;
    Default_Set_Value_Property: string;
    Default_Get_Value_Property: string;
    Default_Value_Changed_Event: string;
    createFromData(data: ISmartObjectItem): void;
    createFromSerialisedData(serialisedData: string): void;
}

export interface ICustomControlSettingSmartObject {
    Name: string;
    Description: string;
    IsRequired: boolean;
    Type: "number" | "boolean" | "text" | "json" | "html" | "date" | "datetime";
    DefaultValue: string;
}